[Refactor] : remove feature delete schedule in home_router
Git : #33
