+++ eb_fast_api/domain/home/tests/home_feature_test.py
@@ -182,64 +182,3 @@ def test_get_path_info(
     finally:
         home_MockScheduleCRUD.dropTable(userEmail=email)
         home_MockPathCRUD.dropTable(user_email=email)
-
-
-def test_delete_schedule(
-    home_MockSession,
-    home_MockPathCRUD,
-    home_MockScheduleCRUD,
-    home_MockUserCRUD,
-):
-    try:
-        # given
-        email = "email"
-        password = "password"
-        refreshToken = "refreshToken"
-        nickName = "nickName"
-        registerInfo = RegisterInfo(
-            nickName=nickName,
-            email=email,
-            password=password,
-        )
-        user = registerInfo.toUser(refreshToken=refreshToken)
-        home_MockUserCRUD.create(user=user)
-
-        path_info = PathInfo.mock()
-        schedule_id = str(uuid4())
-        schedule = Schedule.mock(id=schedule_id)
-        home_MockScheduleCRUD.create(
-            userEmail=user.email,
-            schedule=schedule,
-        )
-        path = path_info.to_path(id=schedule_id)
-        home_MockPathCRUD.create(
-            user_email=user.email,
-            path=path,
-        )
-
-        # when
-        home_feature.delete_schedule(
-            session=home_MockSession,
-            user_email=user.email,
-            schedule_id=schedule_id,
-        )
-
-        # then
-        try:
-            home_MockScheduleCRUD.read(user_email=user.email, schedule_id=schedule_id)
-            raise Exception("test_delete_schedule_fail")
-        except:
-            pass
-
-        fetched_path_dict = home_MockPathCRUD.read(
-            user_email=user.email,
-            path_id=schedule_id,
-        )
-        if fetched_path_dict:
-            raise Exception("test_delete_schedule_fail")
-        else:
-            return
-    # delete schedule table
-    finally:
-        home_MockScheduleCRUD.dropTable(userEmail=email)
-        home_MockPathCRUD.dropTable(user_email=email)
