+++ eb_fast_api/domain/schedule/tests/schedule_feature_test.py
@@ -5,6 +5,7 @@ from eb_fast_api.domain.schema.sources.schemas import (
     PlaceInfo,
     PathInfo,
 )
+from eb_fast_api.database.sources.model.models import Schedule
 from datetime import datetime
 from uuid import uuid4
 
@@ -172,6 +173,69 @@ def test_update_schedule(
         expect_schedule = to_update.toSchedule()
         assert expect_schedule.to_dict() == fetched_schedule_dict
 
+    # delete schedule table
+    finally:
+        schedule_MockScheduleCRUD.dropTable(userEmail=email)
+        schedule_MockPathCRUD.dropTable(user_email=email)
+
+
+def test_delete_schedule(
+    schedule_MockSession,
+    schedule_MockUserCRUD,
+    schedule_MockScheduleCRUD,
+    schedule_MockPathCRUD,
+):
+    try:
+        # given
+        email = "email"
+        password = "password"
+        refreshToken = "refreshToken"
+        nickName = "nickName"
+        registerInfo = RegisterInfo(
+            nickName=nickName,
+            email=email,
+            password=password,
+        )
+        user = registerInfo.toUser(refreshToken=refreshToken)
+        schedule_MockUserCRUD.create(user=user)
+
+        path_info = PathInfo.mock()
+        schedule_id = str(uuid4())
+        schedule = Schedule.mock(id=schedule_id)
+        schedule_MockScheduleCRUD.create(
+            userEmail=user.email,
+            schedule=schedule,
+        )
+        path = path_info.to_path(id=schedule_id)
+        schedule_MockPathCRUD.create(
+            user_email=user.email,
+            path=path,
+        )
+
+        # when
+        schedule_feature.delete_schedule(
+            session=schedule_MockSession,
+            user_email=user.email,
+            schedule_id=schedule_id,
+        )
+
+        # then
+        try:
+            schedule_MockScheduleCRUD.read(
+                user_email=user.email, schedule_id=schedule_id
+            )
+            raise Exception("test_delete_schedule_fail")
+        except:
+            pass
+
+        fetched_path_dict = schedule_MockPathCRUD.read(
+            user_email=user.email,
+            path_id=schedule_id,
+        )
+        if fetched_path_dict:
+            raise Exception("test_delete_schedule_fail")
+        else:
+            return
     # delete schedule table
     finally:
         schedule_MockScheduleCRUD.dropTable(userEmail=email)
